+++ tests/test_edge_current_behavior.py
@@ -1,42 +1,5 @@
-import importlib, importlib.util, sys, pathlib
-
-def _import_by_filename(module_name: str, filename: str):
-    root = pathlib.Path(__file__).resolve().parents[1]
-    path = next(root.rglob(filename), None)
-    if not path:
-        raise ImportError(f"Cannot find {filename} from {root}")
-    spec = importlib.util.spec_from_file_location(module_name, str(path))
-    mod = importlib.util.module_from_spec(spec)
-    spec.loader.exec_module(mod)  # type: ignore
-    sys.modules[module_name] = mod
-    return mod
-
-def _ensure_models_then_app():
-    try:
-        importlib.import_module("models")
-    except Exception:
-        _import_by_filename("models","models.py")
-    try:
-        return importlib.import_module("app")
-    except Exception:
-        return _import_by_filename("app","app.py")
-
-import pytest
-
-@pytest.fixture(scope="module")
-def app_mod():
-    return _ensure_models_then_app()
-
-@pytest.fixture(scope="module")
-def client(app_mod):
-    app = getattr(app_mod, "app")
-    app.config.update(TESTING=True, SECRET_KEY="test")
-    return app.test_client()
-
 def test_checkout_failure_when_card_declines(client):
-    # add item
     client.post("/add-to-cart", data={"title":"1984", "quantity":1}, follow_redirects=True)
-    # attempt checkout with known-declining card (ends with 1111)
     r = client.post("/process-checkout", data={
         "name": "Bob",
         "email": "bob@example.com",
@@ -44,21 +7,16 @@ def test_checkout_failure_when_card_declines(client):
         "city": "Testville",
         "zip_code": "99999",
         "payment_method": "credit_card",
-        "card_number": "4111111111111111",  # this should fail per PaymentGateway
+        "card_number": "4111111111111111",
         "expiry_date": "12/30",
         "cvv": "111",
         "discount_code": "",
     }, follow_redirects=True)
-    # After failure, app redirects back to /checkout and flashes error
     assert r.status_code in (200, 302)
 
-def test_update_cart_zero_quantity_current_behavior(client, app_mod):
-    # This test documents current behavior: /update-cart with qty <= 0 flashes "Removed",
-    # but models.Cart.update_quantity sets quantity to 0 (does not delete the item).
-    # We don't fail the test; we just ensure the route works and doesn't 500.
+def test_update_cart_zero_quantity_current_behavior(client):
     client.post("/add-to-cart", data={"title":"Moby Dick", "quantity":1}, follow_redirects=True)
     r = client.post("/update-cart", data={"title":"Moby Dick", "quantity":0}, follow_redirects=True)
     assert r.status_code in (200, 302)
-    # Optionally, ensure we can still render /cart and /checkout won't crash
     r = client.get("/cart")
     assert r.status_code == 200
